+++ src/common/ChatError.ts
@@ -6,8 +6,8 @@ export class ChatError {
    * The error code.
    *
    * See the error code of the iOS or Android platform:
-   * - iOS: {@link EMErrorCode#https://docs-im.easemob.com/ccim/ios/errorcode}
-   * - Android: {@link HyphenateException#errorCode#https://docs-im.easemob.com/ccim/android/errorcode}
+   * - iOS: {@url https://docs-preprod.agora.io/en/agora-chat/agora_chat_error_ios?platform=iOS}
+   * - Android: {@url https://docs-preprod.agora.io/en/agora-chat/agora_chat_error_android?platform=Android}
    */
   code: number;
   /**